mixer: organize by type
