+++ src/lib/mixer/HelicopterMixer/HelicopterMixer.cpp
