+++ src/lib/mixer/MultirotorMixer/geometries/dodeca_bottom_cox.toml
