+++ src/lib/mixer/MultirotorMixer/geometries/dodeca_top_cox.toml
