+++ src/lib/mixer/MultirotorMixer/geometries/hex_cox.toml
