+++ src/lib/mixer/MultirotorMixer/geometries/hex_t.toml
