+++ src/lib/mixer/MultirotorMixer/geometries/octa_cox.toml
