+++ src/lib/mixer/MultirotorMixer/geometries/octa_cox_wide.toml
