+++ src/lib/mixer/MultirotorMixer/geometries/quad_wide.toml
