+++ src/lib/mixer/MultirotorMixer/geometries/quad_y.toml
