+++ src/lib/mixer/MultirotorMixer/geometries/tri_y.toml
