+++ src/lib/mixer/NullMixer/CMakeLists.txt
@@ -31,26 +31,9 @@
 #
 ############################################################################
 
-# required by other mixers
-add_subdirectory(Mixer)
-
-add_subdirectory(HelicopterMixer)
-add_subdirectory(MultirotorMixer)
-add_subdirectory(NullMixer)
-add_subdirectory(SimpleMixer)
-
-add_library(mixer
-	MixerGroup.cpp
-	MixerGroup.hpp
-	load_mixer_file.cpp
-	mixer_load.h
+add_library(NullMixer
+	NullMixer.cpp
+	NullMixer.hpp
 )
-
-target_link_libraries(mixer
-	PRIVATE
-		HelicopterMixer
-		MultirotorMixer
-		NullMixer
-		SimpleMixer
-	)
-add_dependencies(mixer prebuild_targets)
+target_link_libraries(NullMixer PRIVATE Mixer)
+add_dependencies(NullMixer prebuild_targets)
