+++ src/lib/mixer/SimpleMixer/CMakeLists.txt
@@ -31,26 +31,9 @@
 #
 ############################################################################
 
-# required by other mixers
-add_subdirectory(Mixer)
-
-add_subdirectory(HelicopterMixer)
-add_subdirectory(MultirotorMixer)
-add_subdirectory(NullMixer)
-add_subdirectory(SimpleMixer)
-
-add_library(mixer
-	MixerGroup.cpp
-	MixerGroup.hpp
-	load_mixer_file.cpp
-	mixer_load.h
+add_library(SimpleMixer
+	SimpleMixer.cpp
+	SimpleMixer.hpp
 )
-
-target_link_libraries(mixer
-	PRIVATE
-		HelicopterMixer
-		MultirotorMixer
-		NullMixer
-		SimpleMixer
-	)
-add_dependencies(mixer prebuild_targets)
+target_link_libraries(SimpleMixer PRIVATE Mixer)
+add_dependencies(SimpleMixer prebuild_targets)
