+++ src/lib/mixer/SimpleMixer/SimpleMixer.hpp
@@ -33,7 +33,7 @@
 
 #pragma once
 
-#include "Mixer.hpp"
+#include <mixer/Mixer/Mixer.hpp>
 
 /** simple channel scaler */
 struct mixer_scaler_s {
