+++ src/modules/mc_rate_control/RateControl/RateControl.hpp
@@ -42,7 +42,7 @@
 #include <matrix/matrix/math.hpp>
 #include <mathlib/math/filter/LowPassFilter2pVector3f.hpp>
 
-#include <lib/mixer/MultirotorMixer.hpp>
+#include <lib/mixer/MultirotorMixer/MultirotorMixer.hpp>
 #include <uORB/topics/rate_ctrl_status.h>
 
 class RateControl
